Jet Core ------------ * ORM DataModel: Bug fix: An assigned sub-entity that has not yet been saved must not be deleted when the parent entity is deleted.

+++ library/Jet/DataModel/Trait/Delete.php
@@ -63,7 +63,10 @@ trait DataModel_Trait_Delete
 
 				if(is_array($prop)) {
 					foreach($prop as $v) {
-						if( $v instanceof DataModel_Related ) {
+						if(
+							$v instanceof DataModel_Related &&
+							$v->getIsSaved()
+						) { 
 							$v->delete();
 						}
 					}
